Remove RotateAboutObjectCenter
This option was identical to RotateAboutGrabPoint, except for one hand near interaction, where it would behave strangely, maintaining a fixed position from the hand, but rotating as the hand rotates. It is not clear that this behaviour is desirable, and the inclusion of this seems like an unnecessary edge case.